Created a rewind class which extends action class
