+++ includes/exercise_reminder_action.php
@@ -1,0 +1,45 @@
+<?php
+
+    if ( !defined( "ABSPATH" ) ) {
+
+        header( 'HTTP/1.0 404 File Not Found', 404 );
+        exit('Error 404 - Page Not Found');
+
+    }
+
+    require_once 'includes/exercise_action.php';
+
+    class ExerciseReminderAction extends ExerciseAction {
+
+        private $graded;
+        private $length;
+        private $enabled;
+        private $grade_value;
+
+        // getter
+        public function __get( $property ) {
+
+            if ( property_exists( $this, $property ) ) {
+
+                return $this->$property;
+
+            }
+
+        }
+
+        // setter
+        public function __set( $property, $value ) {
+
+            if ( property_exists( $this, $property ) ) {
+
+                $this->$property = $value;
+
+            }
+
+            return $this;
+
+        }
+
+    }
+
+?>
